requirements.txt
pandas
matplotlib
ipympl
opencv-python
nudenet
pillow
numpy